// src/interfaces/details.ts
// Generated by https://quicktype.io

export interface Details {
    html_attributions: any[];
    result:            Result;
    status:            string;
}

export interface Result {
    business_status:        string;
    formatted_phone_number: string;
    name:                   string;
    opening_hours:          OpeningHours;
    vicinity:               string;
}

export interface OpeningHours {
    open_now:     boolean;
    periods:      Period[];
    weekday_text: string[];
}

export interface Period {
    close: Close;
    open:  Close;
}

export interface Close {
    day:  number;
    time: string;
}